:sparkles: Added gender and name inflection to Player model
Added use of GenderService and NameInflectionService in the Player model to include methods for getting player's gender and inflecting player's nickname. Modified load.php to change the way languages are fetched and read. Also, added 'man' and 'woman' functions in modify.php and adjusted the way names are handled in NameInflectionService.